requirements.txt
absl-py==0.7.1
astor==0.8.0
bleach==1.5.0
cffi==1.12.3
cloudpickle==1.1.1
cycler==0.10.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
horovod==0.16.3
html5lib==0.9999999
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.8.0
psutil==5.6.2
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
tensorboard==1.8.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.15.4
